99-SideProject/django3.1test/djangoAPI/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.7.0
djangorestframework==3.11.1
pytz==2021.1
sqlparse==0.4.1
